# Maintain BlockHandle of meta blocks by BlockManager