--- a/src/include/duckdb/storage/block_manager.hpp
+++ b/src/include/duckdb/storage/block_manager.hpp
@@ -65,9 +65,9 @@ public:
 	virtual idx_t FreeBlocks() = 0;
 
 	//! Register a block with the given block id in the base file
-	shared_ptr<BlockHandle> RegisterBlock(block_id_t block_id);
-	//! Register a meta block with the given block id and save the handle to meta_blocks
-	shared_ptr<BlockHandle> RegisterMetaBlock(block_id_t block_id);
+	shared_ptr<BlockHandle> RegisterBlock(block_id_t block_id, bool is_meta_block = false);
+	//! Clear cached handles for meta blocks
+	void ClearMetaBlockHandles();
 	//! Convert an existing in-memory buffer into a persistent disk-backed block
 	shared_ptr<BlockHandle> ConvertToPersistent(block_id_t block_id, shared_ptr<BlockHandle> old_block);
 
@@ -81,7 +81,7 @@ private:
 	mutex blocks_lock;
 	//! A mapping of block id -> BlockHandle
 	unordered_map<block_id_t, weak_ptr<BlockHandle>> blocks;
-	//! A mapping of meta block id -> BlockHandle
+	//! A map to cache the BlockHandles of meta blocks
 	unordered_map<block_id_t, shared_ptr<BlockHandle>> meta_blocks;
 };
 } // namespace duckdb