This PR fix #5698 

A map from `block_id` to `shared_ptr<BlockHandle>` can be maintained by `BlockManager`, making the `BlockHandle` of meta blocks not freed even when `MetaBlockReader` is freed.

To recover a TPC-H(SF=100) checkpoint (27GB), the running time of function `SingleFileStorageManager::LoadDatabase`:
| latest master | this PR  |
|  ----  | ----  |
|  5.4s  | **0.47s** |

There is a speedup of more than 10x. (Note that this experiment is performed when direct_io=false. For direct_io=true, there will be a much greater improvement.)